Refactored some of the Mine related classes to better organize them since they will be undergoing some changes soon.
