+++ prison-mines/src/main/java/tech/mcprison/prison/mines/commands/MinesCommands.java
@@ -42,11 +42,11 @@ import tech.mcprison.prison.internal.block.PrisonBlockTypes;
 import tech.mcprison.prison.mines.PrisonMines;
 import tech.mcprison.prison.mines.data.Block;
 import tech.mcprison.prison.mines.data.Mine;
-import tech.mcprison.prison.mines.data.MineBlockEvent;
 import tech.mcprison.prison.mines.data.MineData;
 import tech.mcprison.prison.mines.data.MineData.MineNotificationMode;
-import tech.mcprison.prison.mines.data.MineLinerBuilder;
-import tech.mcprison.prison.mines.data.MineLinerBuilder.LinerPatterns;
+import tech.mcprison.prison.mines.features.MineBlockEvent;
+import tech.mcprison.prison.mines.features.MineLinerBuilder;
+import tech.mcprison.prison.mines.features.MineLinerBuilder.LinerPatterns;
 import tech.mcprison.prison.mines.data.PrisonSortableResults;
 import tech.mcprison.prison.mines.managers.MineManager;
 import tech.mcprison.prison.mines.managers.MineManager.MineSortOrder;
