+++ prison-mines/src/main/java/tech/mcprison/prison/mines/data/Mine.java
@@ -29,6 +29,7 @@ import tech.mcprison.prison.internal.World;
 import tech.mcprison.prison.internal.block.PrisonBlock;
 import tech.mcprison.prison.mines.MineException;
 import tech.mcprison.prison.mines.PrisonMines;
+import tech.mcprison.prison.mines.features.MineBlockEvent;
 import tech.mcprison.prison.mines.managers.MineManager;
 import tech.mcprison.prison.output.Output;
 import tech.mcprison.prison.selection.Selection;
