+++ prison-mines/src/main/java/tech/mcprison/prison/mines/data/MineReset.java
@@ -13,9 +13,15 @@ import tech.mcprison.prison.internal.Player;
 import tech.mcprison.prison.internal.World;
 import tech.mcprison.prison.internal.block.PrisonBlock;
 import tech.mcprison.prison.mines.PrisonMines;
-import tech.mcprison.prison.mines.data.MineLinerBuilder.LinerPatterns;
 import tech.mcprison.prison.mines.data.MineScheduler.MineJob;
 import tech.mcprison.prison.mines.events.MineResetEvent;
+import tech.mcprison.prison.mines.features.MineLinerBuilder;
+import tech.mcprison.prison.mines.features.MineMover;
+import tech.mcprison.prison.mines.features.MineTargetBlock;
+import tech.mcprison.prison.mines.features.MineTargetBlockKey;
+import tech.mcprison.prison.mines.features.MineTargetPrisonBlock;
+import tech.mcprison.prison.mines.features.MineTracerBuilder;
+import tech.mcprison.prison.mines.features.MineLinerBuilder.LinerPatterns;
 import tech.mcprison.prison.output.Output;
 import tech.mcprison.prison.tasks.PrisonRunnable;
 import tech.mcprison.prison.util.BlockType;
